Fixed some more bugs. Working on notification
