+++ lib/HomeScreen/note_view.dart
@@ -99,13 +99,12 @@ class NoteViewState extends State<NoteView> {
     });
   }
 
-  void handleSelectAllChange(bool? selectAll) {
+  void handleSelectAllChange(bool selectAll) {
     setState(() {
-      selectedItems = List.filled(displayedNotes.length, selectAll ?? false);
+      selectedItems = List.filled(displayedNotes.length, selectAll);
       // List.filled(notesNotifier.value.length, selectAll ?? false);
-      if (selectAll!) {
-        notesKeys = List.filled(
-            displayedNotes.length, (index) => displayedNotes[index].key);
+      if (selectAll) {
+        notesKeys = displayedNotes.map((note) => note.key).toList();
       } else {
         notesKeys = [];
       }
@@ -155,17 +154,6 @@ class NoteViewState extends State<NoteView> {
     }
   }
 
-  void handleLocationOperation(String locationOperation) {
-    if (locationOperation == 'addLocation') {
-      addNewLocation(context);
-      setpreDefinedLocation();
-    }
-
-    if (locationOperation == 'removeLocation') {
-      removeLocation(context, preDefinedLoations);
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     return isLoading
@@ -177,60 +165,42 @@ class NoteViewState extends State<NoteView> {
             body: const Center(child: CircularProgressIndicator()))
         : Scaffold(
             appBar: AppBar(
-              title: const Text("Location Notes"),
-              automaticallyImplyLeading: false,
-              actions: [
-                if (displayedNotes.isNotEmpty) ...[
+                title: const Text("Location Notes"),
+                automaticallyImplyLeading: false,
+                actions: [
                   Row(
                     children: [
-                      PopupMenuButton(
-                        icon: const Icon(Icons.sort),
-                        itemBuilder: (BuildContext context) => <PopupMenuEntry>[
-                          const PopupMenuItem(
-                            value: 'noteTitle',
-                            child: Text('Sort by Note Title'),
-                          ),
-                          if (isLocationServicesAvailable)
+                      if (displayedNotes.isNotEmpty) ...[
+                        PopupMenuButton(
+                          icon: const Icon(Icons.sort),
+                          itemBuilder: (BuildContext context) =>
+                              <PopupMenuEntry>[
                             const PopupMenuItem(
-                              value: 'location',
-                              child: Text('Sort by Location'),
-                            )
-                        ],
-                        onSelected: (selectedOption) {
-                          sortHandler(selectedOption);
-                        },
-                      ),
-                      const SizedBox(width: 4),
-                      if (notesKeys.isNotEmpty)
-                        IconButton(
-                          icon: const Icon(Icons.delete),
-                          onPressed: () {
-                            deleteSelectedNotes(context, notesKeys);
+                              value: 'noteTitle',
+                              child: Text('Sort by Note Title'),
+                            ),
+                            if (isLocationServicesAvailable)
+                              const PopupMenuItem(
+                                value: 'location',
+                                child: Text('Sort by Location'),
+                              )
+                          ],
+                          onSelected: (selectedOption) {
+                            sortHandler(selectedOption);
                           },
                         ),
-                      const SizedBox(width: 4),
-                      PopupMenuButton(
-                        icon: const Icon(Icons.more_vert),
-                        itemBuilder: (BuildContext context) => <PopupMenuEntry>[
-                          const PopupMenuItem(
-                            value: 'addLocation',
-                            child: Text('Add Location'),
+                        const SizedBox(width: 4),
+                        if (notesKeys.isNotEmpty)
+                          IconButton(
+                            icon: const Icon(Icons.delete),
+                            onPressed: () {
+                              deleteSelectedNotes(context, notesKeys);
+                            },
                           ),
-                          if (preDefinedLoations.isNotEmpty)
-                            const PopupMenuItem(
-                              value: 'removeLocation',
-                              child: Text('Delete Location'),
-                            )
-                        ],
-                        onSelected: (selectedOption) {
-                          handleLocationOperation(selectedOption);
-                        },
-                      ),
+                      ],
                     ],
                   )
-                ]
-              ],
-            ),
+                ]),
             body: Column(
               children: [
                 if (fetchedNotes.isNotEmpty) ...[
@@ -270,7 +240,7 @@ class NoteViewState extends State<NoteView> {
                               value: selectedItems
                                   .every((isSelected) => isSelected),
                               onChanged: (value) {
-                                handleSelectAllChange(value);
+                                handleSelectAllChange(value!);
                               },
                             ),
                           ),
@@ -454,137 +424,10 @@ class NoteViewState extends State<NoteView> {
                   //     .refreshNotifier
                   //     .value = true;
                   notesKeys = [];
-                  getNotes();
                   Navigator.of(context).pop();
+                  getNotes();
                 });
               },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void addNewLocation(BuildContext context) {
-    final formKey = GlobalKey<FormState>();
-    TextEditingController locationController = TextEditingController();
-    TextEditingController coordinatesController = TextEditingController();
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Add Location'),
-          content: Form(
-            key: formKey,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                TextFormField(
-                    controller: locationController,
-                    decoration:
-                        const InputDecoration(labelText: 'Location Name'),
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter Location Name';
-                      } else if (preDefinedLoations.isNotEmpty &&
-                          preDefinedLoations.contains(value)) {
-                        return 'Location Name already exists';
-                      }
-                      return null;
-                    }),
-                const SizedBox(height: 10),
-                TextFormField(
-                    controller: coordinatesController,
-                    decoration: const InputDecoration(labelText: 'Coordinates'),
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter coordinates';
-                      }
-                      return null;
-                    }),
-              ],
-            ),
-          ),
-          actions: [
-            TextButton(
-              child: const Text('Add'),
-              onPressed: () {
-                // Perform the add operation here
-                if (formKey.currentState!.validate()) {
-                  String locationName = locationController.text;
-                  String coordinates = coordinatesController.text;
-
-                  addLocation(locationName, coordinates).then((value) {
-                    formKey.currentState!.reset();
-                    // Close the alert dialog
-                    Navigator.of(context).pop();
-                  });
-                }
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void removeLocation(BuildContext context, List<dynamic> locations) {
-    final formKey = GlobalKey<FormState>();
-    dynamic selectedLocation;
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Delete Location'),
-          content: Form(
-            key: formKey,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                DropdownButtonFormField<String>(
-                  icon: const Icon(Icons.arrow_drop_down),
-                  menuMaxHeight: 150,
-                  value: selectedLocation,
-                  onChanged: (value) {
-                    setState(() {
-                      selectedLocation = value;
-                    });
-                  },
-                  items: locations.map((location) {
-                    return DropdownMenuItem<String>(
-                      value: location,
-                      child: Text(location),
-                    );
-                  }).toList(),
-                  decoration: const InputDecoration(
-                    labelText: 'Select Location',
-                    border: OutlineInputBorder(),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please select a location';
-                    }
-                    return null;
-                  },
-                ),
-              ],
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () {
-                if (formKey.currentState!.validate() &&
-                    selectedLocation != null) {
-                  // Perform the delete operation here
-                  // Delete the selected location from the database
-                  deleteLocation(selectedLocation!).then((value) {
-                    setpreDefinedLocation();
-                    Navigator.of(context).pop();
-                  });
-                  // Close the alert dialog
-                }
-              },
-              child: const Text('Delete'),
             ),
           ],
         );
